pyproject.toml
[project]
name = "finetune-recovery"
version = "1.0.0"
description = "Diff Interpretation Tuning (DIT) paper"
requires-python = ">=3.13"
dependencies = [
    "brokenaxes>=0.6.2",
    "datasets>=3.3.2",
    "hf-transfer>=0.1.9",
    "hf-xet>=1.1.0",
    "huggingface-hub>=0.30.2",
    # "inspect-ai==0.3.95",
    # Fork of inspect_ai 0.3.95 that supports temperature=0
    "inspect-ai @ git+https://github.com/ed1d1a8d/inspect_ai.git@7c08da64c2be3ef4931f1589e080aef36e758922",
    "jinja2>=3.1.6",
    "lovely-tensors>=0.1.18",
    "numpydantic>=1.6.11",
    "openai>=1.78.0",
    "peft>=0.14.0",
    "python-dotenv>=1.1.0",
    "scipy>=1.16.1",
    "simple-parsing>=0.1.7",
    "textual<=3.1.1", # 3.2.0 is broken so we hold it back for now
    "torch>=2.6.0",
    "transformers>=4.49.0",
    "wandb>=0.19.11",
]

[project.scripts]
finetune-recovery = "finetune_recovery:main"

[project.entry-points.inspect_ai]
hf_lora = "finetune_recovery.eval.inspect_hf_lora_provider"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
    "gpustat>=1.1.1",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.6",
    "jupyter>=1.1.1",
    "magic-wormhole>=0.18.0",
    "rich>=13.3.3",
    "ruff==0.13.3", # Should match the version in github/workflows/lint.yml
]

[tool.ruff.lint]
select = [
    "I",
    "E",
    "F",
    "UP",
]
ignore = [
    "E501",
]
